Move files in storage/List* into storage/list
