+++ net/encoders/OutputValues.h
@@ -5,7 +5,7 @@
 
 #include "GraphPath.h"
 #include "ID.h"
-#include "ListElementView.h"
+#include "list/ListElementView.h"
 #include "OptionalLike.h"
 #include "TypeUtils.h"
 #include "metadata/PropertyNull.h"
