+++ query/pipeline/processors/UnwindProcessor.cpp
@@ -5,9 +5,9 @@
 
 #include "ExecutionContext.h"
 
-#include "ListBufferTypeTag.h"
-#include "ListElementView.h"
-#include "ListUtils.h"
+#include "list/ListBufferTypeTag.h"
+#include "list/ListElementView.h"
+#include "list/ListUtils.h"
 
 #include "columns/Column.h"
 #include "columns/ColumnVector.h"
