+++ storage/DebugDump.cpp
@@ -4,8 +4,8 @@
 
 #include <range/v3/view/drop.hpp>
 
-#include "ListElementView.h"
-#include "ListUtils.h"
+#include "list/ListElementView.h"
+#include "list/ListUtils.h"
 
 using namespace db;
