+++ storage/columns/AllowedKinds.h
@@ -4,7 +4,7 @@
 #include <optional>
 #include <type_traits>
 
-#include "ListBuffer.h"
+#include "list/ListBuffer.h"
 #include "ID.h"
 #include "versioning/ChangeID.h"
 #include "ColumnOperator.h"
