+++ test/query-test-suite/QueryResultFormatter.cpp
@@ -11,9 +11,9 @@
 #include "EntityList.h"
 #include "GraphPath.h"
 #include "ID.h"
-#include "ListElementView.h"
-#include "ListUtils.h"
-#include "ListView.h"
+#include "list/ListElementView.h"
+#include "list/ListUtils.h"
+#include "list/ListView.h"
 #include "QueryStatus.h"
 #include "columns/AllowedKinds.h"
 #include "columns/ColumnConst.h"
