+++ test/query-test-suite/QueryTestRunner.cpp
@@ -17,8 +17,8 @@
 #include "Graph.h"
 #include "ID.h"
 #include "JsonEncoder.h"
-#include "ListElementView.h"
-#include "ListView.h"
+#include "list/ListElementView.h"
+#include "list/ListView.h"
 #include "PlanGraph.h"
 #include "PlanGraphDebug.h"
 #include "PlanGraphGenerator.h"
